requirements.txt
asgiref==3.6.0
Django==4.2
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.4.1
Markdown==3.4.3
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
zipp==3.15.0
